Another try with add more layer to resource - import * is not look good - importing may be a problem later - *** may find another better way
