+++ omise/api/resources/balance.py
@@ -1,7 +1,8 @@
-from omise import Base, _as_object, _MainResource
+from omise.api import *
+from omise.api.resources import *
 
 
-class Balance(_MainResource, Base):
+class Balance(MainResource, Base):
     """API class representing balance details.
 
     This API class is used for retrieving current balance of the account.
@@ -27,7 +28,7 @@ class Balance(_MainResource, Base):
 
         :rtype: Balance
         """
-        return _as_object(cls._request('get', cls._instance_path()))
+        return as_object(cls._request('get', cls._instance_path()))
 
     def reload(self):
         """Reload the balance details.
