+++ omise/api/resources/link.py
@@ -1,7 +1,8 @@
-from omise import Base, _as_object, LazyCollection, _MainResource
+from omise.api import *
+from omise.api.resources import *
 
 
-class Link(_MainResource, Base):
+class Link(MainResource, Base):
     """API class representing a link.
 
     This API class is used for retrieving and creating a link.
@@ -49,7 +50,7 @@ class Link(_MainResource, Base):
         :param \*\*kwargs: arguments to create a link.
         :rtype: Link
         """
-        return _as_object(
+        return as_object(
             cls._request('post',
                          cls._collection_path(),
                          kwargs))
@@ -63,16 +64,16 @@ class Link(_MainResource, Base):
         :rtype: Link
         """
         if link_id:
-            return _as_object(
+            return as_object(
                 cls._request('get',
                              cls._instance_path(link_id)))
-        return _as_object(cls._request('get', cls._collection_path()))
+        return as_object(cls._request('get', cls._collection_path()))
 
     @classmethod
     def list(cls):
         """Return all links that belongs to your account.
 
-        :rtype: LazyCollection
+        :rtype: omise.api.resources.lazy_collection.LazyCollection
         """
         return LazyCollection(cls._collection_path())
